+++ src/components/BaseThemeOverlay.tsx
@@ -1,6 +1,9 @@
-import React from 'react'
+import React, { useContext, useState } from 'react'
 
+import { PrestoContext } from '../context/PrestoContext'
 import { useAd } from '../hooks/hooks'
+import { State } from '../Player'
+import { usePrestoUiEvent } from '../react'
 import { ControlsVisibilityMode } from '../services/controls'
 
 import { AdCountdown } from './AdCountdown'
@@ -119,6 +122,17 @@ export interface BaseThemeOverlayProps extends BaseComponentProps {
   version?: number
 }
 
+function useIdleState (): boolean {
+  const { player } = useContext(PrestoContext)
+  const [isIdle, setIdle] = useState(
+    player.state === State.Idle || player.state === State.Unset,
+  )
+  usePrestoUiEvent('statechanged', () => {
+    setIdle(player.state === State.Idle || player.state === State.Unset)
+  })
+  return isIdle
+}
+
 /**
  * Base theme overlay.
  */
@@ -131,7 +145,10 @@ export const BaseThemeOverlay = (props: BaseThemeOverlayProps) => {
   const showSeekBarCues = props.showSeekBarCues ?? true
   const hasPauseButton = props.hasPauseButton ?? true
   const hasTime = props.hasTime ?? true
+  const { player } = useContext(PrestoContext)
   const ad = useAd()
+  const isIdle = useIdleState()
+  const isStartButtonVisible = props.startButton && isIdle
 
   const isFullScreen = useIsPlayerFullScreen()
 
@@ -189,9 +206,13 @@ export const BaseThemeOverlay = (props: BaseThemeOverlayProps) => {
 
   const topCompanion = props.renderTopCompanion?.(isFullScreen)
 
+  const onClickFreeSpace = () => {
+    player.playing = !player.playing
+  }
+
   return (
     <div data-testid="pp-ui-basic-theme" className={'pp-ui pp-ui-overlay pp-ui-basic-theme'} style={props.style}>
-      <PlayerControls mode={props.controlsVisibility}>
+      <PlayerControls mode={isStartButtonVisible ? 'never' : props.controlsVisibility}>
         <VerticalBar className={'pp-ui-spacer'}>
 
           {/* Top bar */}
@@ -202,10 +223,10 @@ export const BaseThemeOverlay = (props: BaseThemeOverlayProps) => {
             </div>
           ): null}
 
-          <Spacer/>
+          <Spacer onClick={onClickFreeSpace}/>
 
           {/* Thumbnails */}
-          <HorizontalBar className={'pp-ui-transparent'}>
+          <HorizontalBar className={'pp-ui-transparent'} onClick={onClickFreeSpace}>
             <Thumbnail moveRelativeToParent={true}/>
           </HorizontalBar>
 
@@ -235,10 +256,13 @@ export const BaseThemeOverlay = (props: BaseThemeOverlayProps) => {
                 </ForSize>
               ): null}
 
-              {hasAudioControls ? <MuteButton/>: null}
-
+              <div className="pp-ui-volume-group">
+                {hasAudioControls ? <MuteButton/> : null}
+                <ForSize size="medium">
+                  {hasAudioControls ? <VolumeBar adjustWhileDragging={true}/> : null}
+                </ForSize>
+              </div>
               <ForSize size="medium">
-                {hasAudioControls ? <VolumeBar adjustWhileDragging={true}/> : null}
                 {hasFullScreenButton ? <FullscreenButton/> : null}
               </ForSize>
             </div>